created serverless function using vercel that will act as a proxy to the api for the client, this is to ensure that the api key does not leak out in teh client application, as of now, the serverless fucntion can make a successfull call to the geo locaiton api, but we have not have it returned by use query yet
